refactor: 공고 필터 모듈화 (#56)
