+++ src/main/java/com/example/eumserver/domain/announcement/filter/domain/OccupationClassification.java
@@ -1,4 +1,4 @@
-package com.example.eumserver.domain.announcement.team.domain;
+package com.example.eumserver.domain.announcement.filter.domain;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
