+++ src/main/java/com/example/eumserver/domain/announcement/team/domain/TeamAnnouncement.java
@@ -1,5 +1,6 @@
 package com.example.eumserver.domain.announcement.team.domain;
 
+import com.example.eumserver.domain.announcement.filter.domain.OccupationClassification;
 import com.example.eumserver.domain.announcement.team.dto.TeamAnnouncementUpdateRequest;
 import com.example.eumserver.domain.team.Team;
 import com.example.eumserver.global.dto.TimeStamp;
