+++ src/main/java/com/example/eumserver/domain/announcement/team/dto/TeamAnnouncementUpdateRequest.java
@@ -1,6 +1,6 @@
 package com.example.eumserver.domain.announcement.team.dto;
 
-import com.example.eumserver.domain.announcement.team.domain.OccupationClassification;
+import com.example.eumserver.domain.announcement.filter.domain.OccupationClassification;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import java.time.LocalDateTime;
